+++ packages/atproto/test/src/atproto_test.dart
@@ -15,6 +15,28 @@ import 'package:atproto/src/servers/servers_service.dart';
 import 'package:atproto/src/sync/sync_service.dart';
 
 void main() {
+  group('.session', () {
+    test('fromSession', () {
+      final session = core.Session(
+        did: 'aaaa',
+        handle: 'bbbbb',
+        accessJwt: 'cccccc',
+        refreshJwt: 'ddddddd',
+      );
+
+      final atproto = ATProto.fromSession(session);
+
+      expect(atproto.session != null, isTrue);
+      expect(atproto.session, session);
+    });
+
+    test('anonymous', () {
+      final atproto = ATProto.anonymous();
+
+      expect(atproto.session == null, isTrue);
+    });
+  });
+
   test('.servers', () {
     final service = ATProto.anonymous().servers;